+++ doc/release-notes.md
@@ -9,3 +9,5 @@ Bitcoin ABC version 0.32.12 is now available from:
   possible to open manual connections to the Tor network for example with the
   `addnode` RPC. To mimic the old behavior use `-proxy=` together with
   `-onlynet=` listing all relevant networks except `onion`.
+- With I2P connections, a new, transient address is used for each outbound
+  connection if `-i2pacceptincoming=0`.